refactor ShardingProperties
