+++ sharding-core/src/main/java/io/shardingsphere/core/constant/properties/ShardingProperties.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.core.constant;
+package io.shardingsphere.core.constant.properties;
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Strings;
